Add the results/stats tabs for all users
Users can see the results and stats tabs even if
the data is not published. When there is nothing to
see, a message is shown saying that the results or
stats have not been published yet. The admins
can see the data always, so, to know if they are
published, the same message appears if they aren't.