Add JS functions to disable the button after click
To prevent double click on the button, it is disabled
after clicking it. Then, when the update process
finishes, the on click event must be binded again because
when the button is painted, that event is lost (it's a new
link created with AJAX).

For IE, disable attribute disables the link, but for the rest
of the browsers isn't working. pointer-events, however,
works for all browsers but IE, so, when the button is clicked
the attribute disabled is set to the link, and, through CSS,
all the links with a certain class have the pointer-events
to none. That way, it is working for all the browsers