Add specs and modify existing ones
A new spec has been added to test that the map
div receives the correct data. For some reason,
the map is not been rendered in the test, so
it's not posible to see if the markers are been
rendered correctly.

Some existing test has been modified to work with
the new map of budget heading's header.