Update controllers and models that use 'proposals' word
related_content_controller.rb uses a little translator
to get the model name from the URL. In the URL, the name
is in spanish (because we have translated it), so it gets
'Propuestas'. To get the name it should be singularized
and translated into english, that's why it's done there
with the 'Budget::Investment' class.

related_content.rb model has a RELATIONABLE_MODELS constant
variable that stores all the relationable models in an array.
The proposals model name was in english, so it have been
changed into spanish 'propuestas'.