Do not count delegators as participants in public stats
Delegators are the users that delegated their vote on a forum. The votes from forums have already been migrated to user forums, so the delegators are already beeing counted as voters.

Maintaining this line would unecessarity count these users twice and we are in the process of removing the delegation references, for hopefully a rewrite soon.