Verify that the ballot line does not exist
This commits adds a special validation, to make sure the ballot line does not already exist. Otherwise, if the rake task is executed twice it would raise an exception. This is due to not validating that a ballot line already exists in the current Budget::Ballot::Line validations[1].

We did not add this validation because the creation of ballot lines is done via ajax and so it did not create a visual failure to the user.

We are however, validating that a ballot line for the same investment does not already exist at the database level[2]. And and exception is raised in that case, which should eventually be fixed.

[1] https://github.com/AyuntamientoMadrid/consul/blob/master/app/models/budget/ballot/line.rb#L10
[2] https://github.com/AyuntamientoMadrid/consul/blob/master/db/schema.rb#L150