Add validation for existing represented user ballots
Given the following scenario:
A Forum has created a Ballot and a Represented User of that Forum had also created a Ballot.

We were adding the ballot lines of the Forum to the Represented User's Ballot.

With this commit we are respecting the Ballot of the Represented User as it is, even if the Forum balloted additional Investments.