Migrate internal comments with a large body
Some internal valuation comments exceed the maxium length of 1000 characters[1] in the Comments validations[2].

This comments were not being migrated properly from Spending Proposals to Budget::Investments.

With this commit we are allowing valuation comments to be created even if they exceeed the limit of 1000 characters.

Note that at the database level we are using a text field[3] for the body of a comment, so we should have plenty of space for large comments.

[1] https://github.com/AyuntamientoMadrid/consul/blob/master/db/seeds.rb#L26

[2] https://github.com/AyuntamientoMadrid/consul/blob/master/app/models/comment.rb#L23

[3] https://github.com/AyuntamientoMadrid/consul/blob/master/db/schema.rb#L428