Require depencies for Budget and Budget::Ballot classes
Eventhoug the specs run fine, when running the rake task[1] to migrate spending proposal ballots, we are getting an exception as Rails is not finding the correct `Budget` and `Budget::Ballot` classes.

By requiring these depedencies explicitly in this migration class the `Budget` and `Budget::Ballot` classes are loaded correctly.

[1] https://github.com/AyuntamientoMadrid/consul/pull/1868