Sort investments by cached votes plus physical votes
The admin interface was displaying investments in a weird order because it did not take into account physical votes when ordering proposals.

With this commit we are taking into account physical votes when ordering investments in the admin interface.

`physical_votes` is a deprecated attribute used to store an integer with the number physical votes for an investment, usually collected by paper signature sheets.

Nowadays we have a SignatureSheet class where admins can insert the data form these signature sheets, which get coverted to real votes of the Vote class.