Remove unused (and flaky) card code and its spec
This code was introduced in commit 1cf1126f, but it stopped being used
in commit ac56a458.

The spec was failing sometimes because the method didn't order the
records, but the spec checked the records returned in a certain order.

Furthermore, the method `page` generated a potential conflict with
kaminari's `page method.